Introduce limited formatter customization
The default jackson formatter produces stange looking output. A custom
`PrettyPrinter` (`ConventionalPrettyPrinter`) results in output more typical
for standard tools like jq and npm.

`ConventionalPrettyPrinter` is made the default for `ModifyMojo`. The new
parameter `formatter` allows the user to go back to the jackson formatter if
they want.

Also update the README and bump the version to 1.2.0